Only initialize sticky nav appear on index
