+++ app/contact.html
@@ -33,7 +33,7 @@
   <link rel="stylesheet" href="./style.css">
 
   <!-- JavaScripts -->
-  <script type="text/javascript" src="index.bundle.js" async></script>
+  <script type="text/javascript" src="base.bundle.js" async></script>
 
 </head>
