+++ app/javascripts/hamburger.js
@@ -1,8 +1,9 @@
-window.onload = function () {
+export default function () {
   // HAMBURGER FOR MOBILE
   var navLinkContainers = document.getElementsByClassName("nav-links");
   var closeIcons = document.getElementsByClassName("close-icon");
   var openIcons = document.getElementsByClassName("open-icon");
+  var samePageLinks = document.getElementsByClassName("same-page-link");
 
   function openHamburger() {
     // Show the navlinks
@@ -39,5 +40,9 @@ window.onload = function () {
   for (var i = 0; i < openIcons.length; i++) {
     var icon = openIcons[i];
     icon.addEventListener('click', openHamburger);
+  }
+  for (var i = 0; i < samePageLinks.length; i++) {
+    var link = samePageLinks[i];
+    link.addEventListener('click', closeHamburger);
   }
 }
